Pull apart different storage backend implementations
